pyproject.toml
[tool.poetry]
name = "shift"
version = "0.1.0"
description = ""
authors = ["Nikita Kolesnik <kolesniknikita21@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.97.0"
pytest = "^7.3.2"
jinja2 = "^3.1.2"
sqlalchemy = "^2.0.16"
envparse = "^0.2.0"
isort = "^5.12.0"
black = "^23.3.0"
psycopg2-binary = "^2.9.6"
alembic = "^1.11.1"
asyncpg = "^0.27.0"
greenlet = "^2.0.2"
pydantic = {extras = ["email"], version = "^1.10.9"}
starlette-context = "^0.3.6"
uvicorn = "^0.22.0"
unicorn = "^2.0.1.post1"
starlette-exporter = "^0.16.0"
pyjwt = "2.6.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^11.0.0"}
beanie = "^1.19.2"
httpx = "^0.24.1"
pytest-asyncio = "^0.21.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
